Fix probing mpeg video in program streams
... and probably some other formats as well.  Libav's probe routine
doesn't necessarily return names that match the codec names 
that can be looked up by avcodec_find_decoder_by_name().  So we
have to manually map the names if the lookup fails.  Lookup for
mpeg video started failing with the last Libav bump because
they removed an obsolete "mpegvideo" decoder that we were matching
on.  The correct decoder is "mpeg2video", but probe doesn't return
that string.

Also fix our implementation of the ff_lockmgr callback.  Current
git Libav fails if we don't fix it. So might as well fix it now.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4341 b64f7644-9d1e-0410-96f1-a4d463321fa5