ca_aac: set minimum VBR quality to 1 instead of 0.
That way, (high-low)/granularity divides cleanly by 1.

Old qualities (low to high in increments of granularity):

{ 0, 9, 18, 27, 36, 45, 54, 63, 72, 81, 90, 99, 108, 117, 126 }

(the default, 91, is not in that list)

New qualities (low to high in increments of granularity):

{ 1, 10, 19, 28, 37, 46, 55, 64, 73, 82, 91 (default), 100, 109, 118, 127 }

Actual output is unchanged (ca_aac has quality ranges; a 1-point bump will not change the range for each quality).



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4422 b64f7644-9d1e-0410-96f1-a4d463321fa5