hb_apply_h264_level() update & cleanup.
The caller may specify any width & height now, rather than having to initialize the x264_param_t's fields directly.

Since it's reasonable to expect GUIs to only provide valid levels (if it doesn't, there's a bug that needs fixing), it's OK to abort when the requested level is not valid.



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4944 b64f7644-9d1e-0410-96f1-a4d463321fa5